# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: CAA Compliance Engine — Internal API
  version: 0.1.0
  description: >
    Internal contract between the PHP/Laravel public site & Client Portal
    and the Java/Spring Boot Compliance Engine (Section 2 of
    CAA-Website-Rebuild-Architecture.md). PHP calls this API; it never
    talks to the database directly. Not public-facing — no external auth
    scheme is specified here, only the shape of the contract.
servers:
  - url: http://compliance-engine.internal:8080/api/v1
security:
  - basicAuth: []
...
paths:
...
  /enrollments:
    post:
      tags: [enrollments]
      summary: >
        Create an enrollment. Triggers QuickBooks invoice creation per the
        school type's pricing rules (Section 7) — no capacity check, since
        sessions don't have an enrollment cap (Section 4).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
        '409':
          description: >
            The session is a VR session (Public or Private) and the
            enrollment's client is not flagged vrClient (Michael,
            2026-08-23). A business/commercial gate, not a regulatory
            one — see VrEnrollmentEligibilityService's Javadoc on the
            compliance-engine side for the full reasoning.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /enrollments/{enrollmentId}/roster-status:
    patch:
      tags: [enrollments]
      summary: >
        Update roster status. ARR and CERTIFIED are normally
        system-set (Digital Testing sign-in, passing certification) —
        this endpoint is primarily for staff setting DNC/DNA.
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rosterStatus]
              properties:
                rosterStatus:
                  $ref: '#/components/schemas/RosterStatus'
      responses:
        '200':
          description: Updated

  # ---------------------------------------------------------------------
  # Certifications (Section 3b)
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        Section 3: the entire API requires HTTP Basic Auth as of the
        security addition -- previously wide open. Applies globally
        (SecurityConfig's anyRequest().authenticated()), not just to the
        one path that explicitly lists it below; that explicit listing on
        /auth/me is just to call out its special role in the login flow.
...
  schemas:
...
    RosterStatus:
      type: string
      enum: [ARR, CERTIFIED, DNC, DNA]
...
    Enrollment:
      type: object
      properties:
        id: { type: integer, format: int64 }
        studentId: { type: integer, format: int64 }
        clientId: { type: integer, format: int64 }
        sessionId: { type: integer, format: int64 }
        paymentStatus: { type: string }
        rosterStatus: { $ref: '#/components/schemas/RosterStatus' }
    EnrollmentCreateRequest:
      type: object
      required: [studentId, clientId, sessionId]
      properties:
        studentId: { type: integer, format: int64 }
        clientId: { type: integer, format: int64 }
        sessionId: { type: integer, format: int64 }
...